gemini-pdf-qa-bot/requirements.txt
cat > requirements.txt << 'EOF'
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.13
chromadb==0.4.22
pypdf==3.17.4
gradio==4.13.0
python-dotenv==1.0.0
EOF